requirements.txt
autopep8
django-cors-headers
Django==4.0
djangorestframework==3.13.1
python-dateutil
python-dotenv